% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{invLogit}
\alias{invLogit}
\title{invLogit}
\usage{
invLogit(theta)
}
\arguments{
\item{theta}{A numeric}
}
\value{
Inverse logit of theta
}
\description{
This function returns the inverse logit of the input argument.
}
\details{
This function is an alias for `stats::binomial()$linkinv`
}
\examples{
invLogit(logit(0.3))
invLogit(c(-Inf, 0, Inf))
}
\seealso{
\code{\link[stats]{family}}
}
\author{
Stephan Wojciekowski
}
