% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(p)
}
\arguments{
\item{p}{A numeric in (0, 1)}
}
\value{
logit of p
}
\description{
This function returns the logit of the input argument.
}
\details{
This function is a shorthand for `stats::binomial()$linkfun`
}
\examples{
logit(invLogit(0.3))
logit(c(0, 0.5, 1))
}
\seealso{
\code{\link[stats]{family}}
}
\author{
Stephan Wojciekowski
}
