% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFunctions.R
\name{saveScenarios}
\alias{saveScenarios}
\title{saveScenarios}
\usage{
saveScenarios(scenario_list, save_path = tempdir())
}
\arguments{
\item{scenario_list}{An object of class \code{scenario_list},
as e.g. created with \code{\link[bhmbasket]{simulateScenarios}}
or \code{\link[bhmbasket]{createTrial}}}

\item{save_path}{A string providing the path of the directory in which the scenario data
should be saved, Default: \code{\link[base]{tempdir}}}
}
\value{
A named list of length 2 containing:
\itemize{
\item \code{scenario_numbers}: the scenario numbers of the saved scenarios
\item \code{path}: the directory in which the scenario data were saved
}
}
\description{
This function saves one or more scenario objects of class \code{scenario_list}
to a newly created or existing directory.
}
\details{
Each scenario in \code{scenario_list} is saved separately as an \code{.rds} file with file name
pattern \verb{scenario_data_<scenario_number>.rds}.

If the directory specified by \code{save_path} does not yet exist, it is created automatically.
Existing files with the same scenario number will be overwritten.
}
\examples{
  scenarios_list <- simulateScenarios(
    n_subjects_list     = list(c(10, 20, 30)),
    response_rates_list = list(rep(0.9, 3)),
    n_trials            = 10)

  save_info <- saveScenarios(
    scenario_list = scenarios_list
  )

  scenarios_list_loaded <- loadScenarios(
    scenario_numbers = save_info$scenario_numbers,
    load_path        = save_info$path)
}
\seealso{
\code{\link[bhmbasket]{simulateScenarios}}
\code{\link[bhmbasket]{loadScenarios}}
\code{\link[base]{saveRDS}}
}
\author{
Stephan Wojciekowski
}
