% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorFunctions.R
\name{setPriorParametersStratified}
\alias{setPriorParametersStratified}
\title{setPriorParametersStratified}
\usage{
setPriorParametersStratified(a_j, b_j)
}
\arguments{
\item{a_j}{A vector of positive numerics for \eqn{\alpha}}

\item{b_j}{A vector of positive numerics for \eqn{\beta}}
}
\value{
A list with prior parameters of class \code{prior_parameters_list}
}
\description{
This function sets prior parameters for the binary-endpoint analysis
method \code{"stratified"} for use in \code{\link[bhmbasket]{performAnalyses}}.
}
\details{
The method \code{"stratified"} is a beta-binomial model for binary endpoints that
assesses each cohort individually. The prior parameters are the shape parameters
of the beta prior distributions.

This function is intended for binary endpoints only. For continuous endpoints,
use \code{\link[bhmbasket]{setPriorParametersNormal}}.
}
\examples{
 prior_parameters_stratified <- setPriorParametersStratified(c(1, 2), c(3, 4))
}
\seealso{
\code{\link[bhmbasket]{performAnalyses}}
\code{\link[bhmbasket]{getPriorParameters}}
\code{\link[bhmbasket]{combinePriorParameters}}
\code{\link[bhmbasket]{setPriorParametersBerry}}
\code{\link[bhmbasket]{setPriorParametersExNex}}
\code{\link[bhmbasket]{setPriorParametersExNexAdj}}
\code{\link[bhmbasket]{setPriorParametersPooled}}
\code{\link[bhmbasket]{setPriorParametersStratifiedMix}}
\code{\link[bhmbasket]{setPriorParametersNormal}}
\code{\link[bhmbasket]{getMuVar}}
}
\author{
Stephan Wojciekowski
}
