% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFunctions.R
\name{createTrial}
\alias{createTrial}
\title{createTrial}
\usage{
createTrial(
  n_subjects,
  n_responders = NULL,
  means = NULL,
  sds = NULL,
  endpoint = c("binary", "normal")
)
}
\arguments{
\item{n_subjects}{A vector of integers for the number of subjects in the trial outcome}

\item{n_responders}{A vector of integers for the number of responders in the trial outcome.
Used for binary endpoints.}

\item{means}{A numeric vector of cohort means. Used for continuous endpoints.}

\item{sds}{A numeric vector of cohort standard deviations. Used for continuous endpoints.
Historical continuous cohorts can be represented by \code{sds = 0}.}

\item{endpoint}{A string indicating the endpoint type. Must be one of \code{"binary"} or \code{"normal"},
Default: \code{"binary"}}
}
\value{
An object of class \code{scenario_list} with the scenario data for a single trial outcome.
}
\description{
This function creates an object of class \code{scenario_list}
for a single trial outcome, which can subsequently be analyzed with other functions of
\code{bhmbasket}, e.g. \code{\link[bhmbasket]{performAnalyses}}
}
\details{
For binary endpoints, this function is a wrapper for \code{\link[bhmbasket]{simulateScenarios}}
using \code{response_rates_list}.
For continuous endpoints, it is a wrapper using \code{means_list} and \code{sds_list}.
}
\examples{
## Binary endpoint example
trial_outcome_bin <- createTrial(
  n_subjects   = c(10, 20, 30, 40),
  n_responders = c(1, 2, 3, 4)
)

## Normal endpoint example
trial_outcome_norm <- createTrial(
  n_subjects = c(15, 15, 20, 20),
  means      = c(5.1, 5.4, 6.0, 6.3),
  sds        = c(1.2, 1.1, 1.5, 1.4),
  endpoint   = "normal"
)
}
\seealso{
\code{\link[bhmbasket]{simulateScenarios}}
\code{\link[bhmbasket]{performAnalyses}}
}
\author{
Stephan Wojciekowski, Tathagata Chattopadhyay
}
