% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{loadAnalyses}
\alias{loadAnalyses}
\title{loadAnalyses}
\usage{
loadAnalyses(
  scenario_numbers,
  analysis_numbers = rep(1, length(scenario_numbers)),
  load_path = tempdir()
)
}
\arguments{
\item{scenario_numbers}{A (vector of) positive integer(s) for the scenario number(s)}

\item{analysis_numbers}{A (vector of) positive integer(s) for the analysis number(s),
Default: \code{rep(1, length(scenario_numbers))}}

\item{load_path}{A string providing a path where the scenarios are being stored,
Default: \code{\link[base]{tempdir}}}
}
\value{
Returns an object of class \code{analysis_list}
}
\description{
This function loads one or more analyses previously saved with
\code{\link[bhmbasket]{saveAnalyses}}.
}
\seealso{
\code{\link[bhmbasket]{performAnalyses}}
\code{\link[bhmbasket]{saveAnalyses}}
\code{\link[base]{tempdir}}
}
\author{
Stephan Wojciekowski
}
