% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorFunctions.R
\name{setPriorParametersStratifiedMix}
\alias{setPriorParametersStratifiedMix}
\title{setPriorParametersStratifiedMix}
\usage{
setPriorParametersStratifiedMix(w, a_j, b_j)
}
\arguments{
\item{w}{A numeric vector of mixture weights in \eqn{[0,1]} summing to 1.}

\item{a_j}{A positive numeric matrix of beta shape parameters \eqn{\alpha_j},
with rows per mixture component and columns per cohort.}

\item{b_j}{A positive numeric matrix of beta shape parameters \eqn{\beta_j},
with rows per mixture component and columns per cohort.}
}
\value{
A list with prior parameters of class \code{prior_parameters_list}
}
\description{
This function sets prior parameters for the analysis method \code{"stratified_mix"}
for use in \code{\link[bhmbasket]{performAnalyses}}.
}
\details{
The method \code{"stratified_mix"} is a beta-binomial model that assesses each cohort
individually with a finite mixture beta prior.
See also the R package \code{RBesT}.
}
\examples{
prior_parameters_stratified_mix <- setPriorParametersStratifiedMix(
  w   = c(0.8, 0.2),
  a_j = rbind(c(2, 3), c(1, 1)),
  b_j = rbind(c(8, 7), c(1, 1))
)
}
\seealso{
\code{\link[bhmbasket]{performAnalyses}}
\code{\link[bhmbasket]{getPriorParameters}}
\code{\link[bhmbasket]{combinePriorParameters}}
\code{\link[bhmbasket]{setPriorParametersStratified}}
}
\author{
Stephan Wojciekowski
}
